Crate sync

sync-rs: revamping std::sync

The purpose of this crate is to provide a prototype implementation of a revamp'd std::sync API. It strives to maintain the high-level types while bringing their implementations much closer to the corresponding system primitives where available.

Abstraction layer

The crate id structured into a two primary modes:

The sys module contains 0-cost, very unsafe, raw bindings to the system primitives. The behavior of these primitives may vary slightly across platforms and are generally considered too unsafe to use. It is highly recommended to use the safe primitives at the top level instead.

The crate root has a number of types exported which are all safe to use and provide alternatives to the sys module. These primitives all provide safety features such as poisoning and RAII guards. Types like Mutex and RWLock also provide the ability to contain the data they are protecting.

Poisoning

The Mutex and RWLock types in this module implement a strategy referred to as poisoning in order to prevent access to possibly invalid data. If a thread panics with write-access to one of these two locks. then all future accesses to the lock will panic immediately.

Static initialization

This crate supports a number of statically initialized primitives for use in setting up C libraries, for example. Types which are normally not statically initialized have a Static-prefix type to use (StaticMutex, StaticCondvar, StaticRWLock). A form of one-time initialization (Once) is also provided. All types have a *_INIT constant which may be used to initialize the primitive.

Custom primitives

Using the system-provided mutexes, condition variables, and rwlocks, this crate also builds abstractions such as Once, Semaphore, and Barrier which do not bind to the corresponding system abstraction if one is available.

Modules

atomic
Atomic types

sys
Bindings to system primitives

Structs

Arc
An atomically reference counted wrapper for shared state.

Barrier
A barrier enables multiple tasks to synchronize the beginning of some computation.

Condvar
A Condition Variable

Future
A type encapsulating the result of a computation which may not be complete

Mutex
A mutual exclusion primitive useful for protecting shared data

MutexGuard
An RAII implementation of a "scoped lock" of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

Once
A synchronization primitive which can be used to run a one-time global initialization. Useful for one-time initialization for FFI or related functionality. This type can only be constructed with the ONCE_INIT value.

RWLock
A reader-writer lock

RWLockReadGuard
RAII structure used to release the shared read access of a lock when dropped.

RWLockWriteGuard
RAII structure used to release the exclusive write access of a lock when dropped.

Semaphore
A counting, blocking, semaphore.

SemaphoreGuard
An RAII guard which will release a resource acquired from a semaphore when dropped.

StaticCondvar
Statically allocated condition variables.

StaticMutex
The static mutex type is provided to allow for static allocation of mutexes.

StaticMutexGuard
An RAII implementation of a "scoped lock" of a static mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

StaticRWLock
Structure representing a staticaly allocated RWLock.

StaticRWLockReadGuard
RAII structure used to release the shared read access of a lock when dropped.

StaticRWLockWriteGuard
RAII structure used to release the exclusive write access of a lock when dropped.

TaskPool
A task pool used to execute functions in parallel.

Weak
A weak pointer to an Arc.

Constants

CONDVAR_INIT
Constant initializer for a statically allocated condition variable.

MUTEX_INIT
Static initialization of a mutex. This constant can be used to initialize other mutex constants.

ONCE_INIT
Initialization value for static Once values.

RWLOCK_INIT
Constant initialization for a statically-initialized rwlock.

Traits

AsMutexGuard
A trait for vaules which can be passed to the waiting methods of condition variables. This is implemented by the mutex guards in this module.