Module sync::atomicExperimental

Atomic types

Atomic types provide primitive shared-memory communication between threads, and are the building blocks of other concurrent types.

This module defines atomic versions of a select number of primitive types, including AtomicBool, AtomicInt, AtomicUint, and AtomicOption. Atomic types present operations that, when used correctly, synchronize updates between threads.

Each method takes an Ordering which represents the strength of the memory barrier for that operation. These orderings are the same as C++11 atomic orderings.

Atomic variables are safe to share between threads (they implement Sync) but they do not themselves provide the mechanism for sharing. The most common way to share an atomic variable is to put it into an Arc (an atomically-reference-counted shared pointer).

Most atomic types may be stored in static variables, initialized using the provided static initializers like INIT_ATOMIC_BOOL. Atomic statics are often used for lazy global initialization.

Examples

A simple spinlock:

use std::sync::Arc;
use std::sync::atomic::{AtomicUint, SeqCst};

fn main() {
    let spinlock = Arc::new(AtomicUint::new(1));

    let spinlock_clone = spinlock.clone();
    spawn(proc() {
        spinlock_clone.store(0, SeqCst);
    });

    // Wait for the other task to release the lock
    while spinlock.load(SeqCst) != 0 {}
}

Transferring a heap object with AtomicOption:

use std::sync::Arc;
use std::sync::atomic::{AtomicOption, SeqCst};

fn main() {
    struct BigObject;

    let shared_big_object = Arc::new(AtomicOption::empty());

    let shared_big_object_clone = shared_big_object.clone();
    spawn(proc() {
        let unwrapped_big_object = shared_big_object_clone.take(SeqCst);
        if unwrapped_big_object.is_some() {
            println!("got a big object from another task");
        } else {
            println!("other task hasn't sent big object yet");
        }
    });

    shared_big_object.swap(box BigObject, SeqCst);
}

Keep a global count of live tasks:

use std::sync::atomic::{AtomicUint, SeqCst, INIT_ATOMIC_UINT};

static GLOBAL_TASK_COUNT: AtomicUint = INIT_ATOMIC_UINT;

let old_task_count = GLOBAL_TASK_COUNT.fetch_add(1, SeqCst);
println!("live tasks: {}", old_task_count + 1);

Structs

AtomicBool
A boolean type which can be safely shared between threads.

AtomicInt
A signed integer type which can be safely shared between threads.

AtomicOption
An atomic, nullable unique pointer

AtomicPtr
A raw pointer type which can be safely shared between threads.

AtomicUint
An unsigned integer type which can be safely shared between threads.

Enums

Ordering
Atomic memory orderings

Functions

fence
An atomic fence.