Struct sync::ArcStable

pub struct Arc<T> {
    // some fields omitted
}

An atomically reference counted wrapper for shared state.

Example

In this example, a large vector of floats is shared between several tasks. With simple pipes, without Arc, a copy would have to be made for each task.

use std::sync::Arc;

fn main() {
    let numbers = Vec::from_fn(100, |i| i as f32);
    let shared_numbers = Arc::new(numbers);

    for _ in range(0u, 10) {
        let child_numbers = shared_numbers.clone();

        spawn(proc() {
            let local_numbers = child_numbers.as_slice();

            // Work with the local numbers
        });
    }
}

Methods

impl<T: Send + Sync> Arc<T>

fn new(data: T) -> Arc<T>

Creates an atomically reference counted wrapper.

fn downgrade(&self) -> Weak<T>

Downgrades a strong pointer to a weak pointer.

Weak pointers will not keep the data alive. Once all strong references to the underlying data have been dropped, the data itself will be destroyed.

impl<T: Send + Sync + Clone> Arc<T>

fn make_unique(&mut self) -> &mut T

Acquires a mutable pointer to the inner contents by guaranteeing that the reference count is one (no sharing is possible).

This is also referred to as a copy-on-write operation because the inner data is cloned if the reference count is greater than one.