Struct sync::Barrier

pub struct Barrier {
    // some fields omitted
}

A barrier enables multiple tasks to synchronize the beginning of some computation.

use std::sync::Arc;
use sync::Barrier;

let barrier = Arc::new(Barrier::new(10));
for _ in range(0u, 10) {
    let c = barrier.clone();
    // The same messages will be printed together.
    // You will NOT see any interleaving.
    spawn(proc() {
        println!("before wait");
        c.wait();
        println!("after wait");
    });
}

Methods

impl Barrier

fn new(n: uint) -> Barrier

Create a new barrier that can block a given number of threads.

A barrier will block n-1 threads which call wait and then wake up all threads at once when the nth thread calls wait.

fn wait(&self)

Block the current thread until all tasks has rendezvoused here.

Barriers are re-usable after all tasks have rendezvoused once, and can be used continuously.