Struct sync::Condvar

pub struct Condvar {
    // some fields omitted
}

A Condition Variable

Condition variables represent the ability to block a thread such that it consumes no CPU time while waiting for an event to occur. Condition variables are typically associated with a boolean predicate (a condition) and a mutex. The predicate is always verified inside of the mutex before determining that thread must block.

Functions in this module will block the current thread of execution and are bindings to system-provided condition variables where possible. Note that this module places one additional restriction over the system condition variables: each condvar can be used with precisely one mutex at runtime. Any attempt to use multiple mutexes on the same condition variable will result in a runtime panic. If this is not desired, then the unsafe primitives in sys do not have this restriction.

Example

use std::sync::Arc;
use sync::{Mutex, Condvar};

let pair = Arc::new((Mutex::new(false), Condvar::new()));
let pair2 = pair.clone();

// Inside of our lock, spawn a new thread, and then wait for it to start
spawn(proc() {
    let &(ref lock, ref cvar) = &*pair2;
    let mut started = lock.lock();
    *started = true;
    cvar.notify_one();
});

// wait for the thread to start up
let &(ref lock, ref cvar) = &*pair;
let started = lock.lock();
while !*started {
    cvar.wait(&started);
}

Methods

impl Condvar

fn new() -> Condvar

Creates a new condition variable which is ready to be waited on and notified.

fn wait<T: AsMutexGuard>(&self, mutex_guard: &T)

Block the current thread until this condition variable receives a notification.

This function will atomically unlock the mutex specified (represented by guard) and block the current thread. This means that any calls to notify_*() which happen logically after the mutex is unlocked are candidates to wake this thread up. When this function call returns, the lock specified will have been re-acquired.

Note that this function is susceptible to spurious wakeups. Condition variables normally have a boolean predicate associated with them, and the predicate must always be checked each time this function returns to protect against spurious wakeups.

Panics

This function will panic!() if it is used with more than one mutex over time. Each condition variable is dynamically bound to exactly one mutex to ensure defined behavior across platforms. If this functionality is not desired, then unsafe primitives in sys are provided.

fn wait_timeout<T: AsMutexGuard>(&self, mutex_guard: &T, dur: Duration) -> bool

Wait on this condition variable for a notification, timing out after a specified duration.

The semantics of this function are equivalent to wait() except that the thread will be blocked for no longer than dur. If the wait timed out, then false will be returned. Otherwise if a notification was received then true will be returned.

Like wait, the lock specified will be re-acquired when this function returns, regardless of whether the timeout elapsed or not.

fn notify_one(&self)

Wake up one blocked thread on this condvar.

If there is a blocked thread on this condition variable, then it will be woken up from its call to wait or wait_timeout. Calls to notify_one are not buffered in any way.

To wake up all threads, see notify_one().

fn notify_all(&self)

Wake up all blocked threads on this condvar.

This method will ensure that any current waiters on the condition variable are awoken. Calls to notify_all() are not buffered in any way.

To wake up only one thread, see notify_one().