Struct sync::MutexGuard

pub struct MutexGuard<'a, T: 'a> {
    // some fields omitted
}

An RAII implementation of a "scoped lock" of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

The data protected by the mutex can be access through this guard via its Deref and DerefMut implementations