Struct sync::RWLock

pub struct RWLock<T> {
    // some fields omitted
}

A reader-writer lock

This type of lock allows a number of readers or at most one writer at any point in time. The write portion of this lock typically allows modification of the underlying data (exclusive access) and the read portion of this lock typically allows for read-only access (shared access).

The type parameter T represents the data that this lock protects. It is required that T satisfies Send to be shared across tasks and Sync to allow concurrent access through readers. The RAII guards returned from the locking methods implement Deref (and Deref mut for the write methods) to allow access to the contained of the lock.

RWLocks, like Mutexes, will become poisoned on panics. Note, however, that an RWLock may only be poisoned if a panic occurs while it is locked exclusively (write mode). If a panic occurs in any reader, then the lock will not be poisoned.

Example

use sync::RWLock;

let lock = RWLock::new(5i);

// many reader locks can be held at once
{
    let r1 = lock.read();
    let r2 = lock.read();
    assert_eq!(*r1, 5);
    assert_eq!(*r2, 5);
} // read locks are dropped at this point

// only one write lock may be held, however
{
    let mut w = lock.write();
    *w += 1;
    assert_eq!(*w, 6);
} // write lock is dropped here

Methods

impl<T: Send + Sync> RWLock<T>

fn new(t: T) -> RWLock<T>

Creates a new instance of an RWLock which is unlocked and read to go.

fn read(&self) -> RWLockReadGuard<T>

Locks this rwlock with shared read access, blocking the current thread until it can be acquired.

The calling thread will be blocked until there are no more writers available. There may be other readers currently inside the lock when this method returns.

Returns an RAII guard which will release this thread's shared access once it is dropped.

Panics

This function will panic if the RWLock is poisoned. An RWLock is poisoned whenever a writer panics while holding an exclusive lock. The panic will occur immediately after the lock has been acquired.

fn try_read(&self) -> Option<RWLockReadGuard<T>>

Attempt to acquire this lock with shared read access.

This function will never block and will return immediately if read would otherwise succeed. Returns Some of an RAII guard which will release the shared access of this thread when dropped, or None if the access could not be granted.

Panics

This function will panic if the RWLock is poisoned. An RWLock is poisoned whenever a writer panics while holding an exclusive lock. A panic will only occur if the lock is acquired.

fn write(&self) -> RWLockWriteGuard<T>

Lock this rwlock with exclusive write access, blocking the current thread until it can be acquired.

This function will not return while other writers or other readers currently have access to the lock.

Returns an RAII guard which will drop the write access of this rwlock when dropped.

Panics

This function will panic if the RWLock is poisoned. An RWLock is poisoned whenever a writer panics while holding an exclusive lock. The panic will occur when the lock is acquired.

fn try_write(&self) -> Option<RWLockWriteGuard<T>>

Attempt to lock this rwlock with exclusive write access.

This function does not ever block, and it will return None if a call to write would otherwise block. If successful, an RAII guard is returned.

Panics

This function will panic if the RWLock is poisoned. An RWLock is poisoned whenever a writer panics while holding an exclusive lock. A panic will only occur if the lock is acquired.