Struct sync::StaticCondvar

pub struct StaticCondvar {
    // some fields omitted
}

Statically allocated condition variables.

This structure is identical to Condvar except that it is suitable for use in static initializers for other structures.

Example

use sync::{StaticCondvar, CONDVAR_INIT};

static CVAR: StaticCondvar = CONDVAR_INIT;

Methods

impl StaticCondvar

fn wait<T: AsMutexGuard>(&'static self, mutex_guard: &T)

Block the current thread until this condition variable receives a notification.

See Condvar::wait.

fn wait_timeout<T: AsMutexGuard>(&self, mutex_guard: &T, dur: Duration) -> bool

Wait on this condition variable for a notification, timing out after a specified duration.

See Condvar::wait_timeout.

fn notify_one(&'static self)

Wake up one blocked thread on this condvar.

See Condvar::notify_one.

fn notify_all(&'static self)

Wake up all blocked threads on this condvar.

See Condvar::notify_all.

unsafe fn destroy(&'static self)

Deallocate all resources associated with this static condvar.

This method is unsafe to call as there is no guarantee that there are no active users of the condvar, and this also doesn't prevent any future users of the condvar. This method is required to be called to not leak memory on all platforms.