Struct sync::StaticRWLock

pub struct StaticRWLock {
    // some fields omitted
}

Structure representing a staticaly allocated RWLock.

This structure is intended to be used inside of a static and will provide automatic global access as well as lazy initialization. The internal resources of this RWLock, however, must be manually deallocated.

Example

use sync::{StaticRWLock, RWLOCK_INIT};

static LOCK: StaticRWLock = RWLOCK_INIT;

{
    let _g = LOCK.read();
    // ... shared read access
}
{
    let _g = LOCK.write();
    // ... exclusive write access
}
unsafe { LOCK.destroy() } // free all resources

Methods

impl StaticRWLock

fn read(&'static self) -> StaticRWLockReadGuard

Locks this rwlock with shared read access, blocking the current thread until it can be acquired.

See RWLock::read.

fn try_read(&'static self) -> Option<StaticRWLockReadGuard>

Attempt to acquire this lock with shared read access.

See RWLock::try_read.

fn write(&'static self) -> StaticRWLockWriteGuard

Lock this rwlock with exclusive write access, blocking the current thread until it can be acquired.

See RWLock::write.

fn try_write(&'static self) -> Option<StaticRWLockWriteGuard>

Attempt to lock this rwlock with exclusive write access.

See RWLock::try_write.

unsafe fn destroy(&'static self)

Deallocate all resources associated with this static lock.

This method is unsafe to call as there is no guarantee that there are no active users of the lock, and this also doesn't prevent any future users of this lock. This method is required to be called to not leak memory on all platforms.