Struct sync::TaskPoolExperimental

pub struct TaskPool {
    // some fields omitted
}

A task pool used to execute functions in parallel.

Spawns n worker tasks and replenishes the pool if any worker tasks panic.

Example

let pool = TaskPool::new(4u);

let (tx, rx) = channel();
for _ in range(0, 8u) {
    let tx = tx.clone();
    pool.execute(proc() {
        tx.send(1u);
    });
}

assert_eq!(rx.iter().take(8u).sum(), 8u);

Methods

impl TaskPool

fn new(tasks: uint) -> TaskPool

Spawns a new task pool with tasks tasks.

Panics

This function will panic if tasks is 0.

fn execute(&self, job: proc(): 'static + Send)

Executes the function job on a task in the pool.