Module sync::sys

Bindings to system primitives

This module contains bindings to the OS-provided synchronization primitives. All functions and methods in this module are unsafe as the goal is to provide a 0-cost abstraction and cross-platform abstraction in this module, not to provide a set of safe primitives to use.

Normal usage should favor the top-level types of this crate instead.

Structs

Condvar
An OS-based condition variable.

Mutex
An OS-based mutual exclusion lock.

RWLock
An OS-based reader-writer lock.

Constants

CONDVAR_INIT
Static initializer for condition variables.

MUTEX_INIT
Constant initializer for statically allocated mutexes.

RWLOCK_INIT
Constant initializer for static RWLocks.