Struct sync::sys::Condvar

pub struct Condvar(_);

An OS-based condition variable.

This structure is the lowest layer possible on top of the OS-provided condition variables. It is consequently entirely unsafe to use. It is recommended to use the safer types at the top level of this crate instead of this type.

Methods

impl Condvar

unsafe fn new() -> Condvar

Creates a new condition variable for use.

Behavior is undefined if the condition variable is moved after it is first used with any of the functions below.

unsafe fn notify_one(&self)

Signal one waiter on this condition variable to wake up.

unsafe fn notify_all(&self)

Awaken all current waiters on this condition variable.

unsafe fn wait(&self, mutex: &Mutex)

Wait for a signal on the specified mutex.

Behavior is undefined if the mutex is not locked by the current thread. Behavior is also undefined if more than one mutex is used concurrently on this condition variable.

unsafe fn wait_timeout(&self, mutex: &Mutex, dur: Duration) -> bool

Wait for a signal on the specified mutex with a timeout duration specified by dur (a relative time into the future).

Behavior is undefined if the mutex is not locked by the current thread. Behavior is also undefined if more than one mutex is used concurrently on this condition variable.

unsafe fn destroy(&self)

Deallocate all resources associated with this condition variable.

Behavior is undefined if there are current or will be future users of this condition variable.