Struct sync::sys::Mutex

pub struct Mutex(_);

An OS-based mutual exclusion lock.

This is the thinnest cross-platform wrapper around OS mutexes. All usage of this mutex is unsafe and it is recommended to instead use the safe wrapper at the top level of the crate instead of this type.

Methods

impl Mutex

unsafe fn new() -> Mutex

Creates a newly initialized mutex.

Behavior is undefined if the mutex is moved after the first method is called on the mutex.

unsafe fn lock(&self)

Lock the mutex blocking the current thread until it is available.

Behavior is undefined if the mutex has been moved between this and any previous function call.

unsafe fn try_lock(&self) -> bool

Attempt to lock the mutex without blocking, returning whether it was successfully acquired or not.

Behavior is undefined if the mutex has been moved between this and any previous function call.

unsafe fn unlock(&self)

Unlock the mutex.

Behavior is undefined if the current thread does not actually hold the mutex.

unsafe fn destroy(&self)

Deallocate all resources associated with this mutex.

Behavior is undefined if there are current or will be future users of this mutex.