Struct sync::sys::RWLock

pub struct RWLock(_);

An OS-based reader-writer lock.

This structure is entirely unsafe and serves as the lowest layer of a cross-platform binding of system rwlocks. It is recommended to use the safer types at the top level of this crate instead of this type.

Methods

impl RWLock

unsafe fn new() -> RWLock

Creates a new instance of an RWLock.

Usage of an RWLock is undefined if it is moved after its first use (any function calls below).

unsafe fn read(&self)

Acquire shared access to the underlying lock, blocking the current thread to do so.

Behavior is undefined if the rwlock has been moved between this and any previous methodo call.

unsafe fn try_read(&self) -> bool

Attempt to acquire shared access to this lock, returning whether it succeeded or not.

This function does not block the current thread.

Behavior is undefined if the rwlock has been moved between this and any previous methodo call.

unsafe fn write(&self)

Acquire write access to the underlying lock, blocking the current thread to do so.

Behavior is undefined if the rwlock has been moved between this and any previous methodo call.

unsafe fn try_write(&self) -> bool

Attempt to acquire exclusive access to this lock, returning whether it succeeded or not.

This function does not block the current thread.

Behavior is undefined if the rwlock has been moved between this and any previous methodo call.

unsafe fn read_unlock(&self)

Unlock previously acquired shared access to this lock.

Behavior is undefined if the current thread does not have shared access.

unsafe fn write_unlock(&self)

Unlock previously acquired exclusive access to this lock.

Behavior is undefined if the current thread does not currently have exclusive access.

unsafe fn destroy(&self)

Destroy OS-related resources with this RWLock.

Behavior is undefined if there are any currently active users of this lock.