Trait sync::AsMutexGuard

pub trait AsMutexGuard {
    unsafe fn as_mutex_guard(&self) -> &StaticMutexGuard;
}

A trait for vaules which can be passed to the waiting methods of condition variables. This is implemented by the mutex guards in this module.

Note that this trait should likely not be implemented manually unless you really know what you're doing.

Required Methods

unsafe fn as_mutex_guard(&self) -> &StaticMutexGuard